gun/pistols: merge Gun_Pistols_Control in TRX
This merges pistol control into TRX and offers the ability to look
while shooting at an enemy as an option in both games.

Resolves #4090.
