+++ docs/tr1/CHANGELOG.md
@@ -23,6 +23,7 @@
 - changed LOS algorithm to TR2+ implementation
 - changed the debug position UI to no longer be hidden in photo mode
 - changed save crystal collision to make them easier to activate
+- changed the unrestricted look mode option to include Lara being able to look freely while shooting an enemy (#4090)
 - fixed missing footstep sound effects when Lara climbs off a ladder and when she finishes a handstand (#4030)
 - fixed a crash in custom levels if a flip effect that expects to act on an item is used in a regular trigger (#4085)
 - fixed a crash if trying to kill an enemy by name but there is no naming definition for that object
