+++ src/libtrx/include/libtrx/game/gun/pistols.h
@@ -2,12 +2,11 @@
 
 #include "../lara/types.h"
 
-extern void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type);
+void Gun_Pistols_Control(LARA_GUN_TYPE weapon_type);
 
 void Gun_Pistols_Draw(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_Undraw(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_Ready(LARA_GUN_TYPE weapon_type);
-void Gun_Pistols_Animate(LARA_GUN_TYPE weapon_type);
 
 void Gun_Pistols_DrawMeshes(LARA_GUN_TYPE weapon_type);
 void Gun_Pistols_UndrawMeshLeft(LARA_GUN_TYPE weapon_type);
