--- src/tr1/game/gun/gun_pistols.h
@@ -1,7 +1,0 @@
-#pragma once
-
-// Private gun routines related to pistols.
-//
-// This refers to the Pistols, but also the Magnums and Uzis.
-
-#include <libtrx/game/gun/pistols.h>
