--- src/tr2/game/gun/gun_pistols.h
@@ -1,5 +1,0 @@
-#pragma once
-
-#include <libtrx/game/gun/pistols.h>
-
-void Gun_Pistols_Animate(LARA_GUN_TYPE weapon_type);
